download readme

+++ SessionManagementAnalyserApp/server-script.js
@@ -309,13 +309,13 @@ function cleanUpRepos(dir) {
     let sub_dirs = fs.readdirSync(dir, { withFileTypes: true }).filter(item => item.isDirectory()).map(item => item.name);
     if (sub_dirs.length === 0) {
         for (let i = 0; i < files.length; i++) {
-            if (!extensions.some(e => files[i].endsWith(e)) && files[i] !== "requirements.txt") {
+            if (!extensions.some(e => files[i].endsWith(e)) && files[i] !== "requirements.txt" && files[i] !== "README.md") {
                 fs.unlinkSync(String(dir) + "/" + String(files[i]))
             }
         }
     } else {
         for (let i = 0; i < files.length; i++) {
-            if (!extensions.some(e => files[i].endsWith(e)) && files[i] !== "requirements.txt") {
+            if (!extensions.some(e => files[i].endsWith(e)) && files[i] !== "requirements.txt" && files[i] !== "README.md") {
                 fs.unlinkSync(String(dir) + "/" + String(files[i]))
             }
         }
@@ -688,7 +688,7 @@ function libraryUsagesGrep() {
     });
 }
 
-// downloadAndExtractRepos('../django_filtered_list_final_v2.csv');
+downloadAndExtractRepos('../django_filtered_list_final_v2.csv');
 // findInterestingRepos("Secure-cookie-attribute", "sf_secure_attribute_session_cookie_manually_disabled.txt", true, 0, Number.MAX_VALUE, './repos_with_interesting_results/9bis - repos_with_manually_disabled_secure_session_cookie_flask_login_final_filtered_merged_list.txt'); // if third parameter is set to true it will look for queries that returned a result, otherwise it will look for queries that didn't return a result
 // findInterestingRepos("HTTPOnly-cookie-attribute", "un_httponly_attribute_session_cookie.txt", true, 0, Number.MAX_VALUE, './repos_with_interesting_results/9bis - repos_with_disabled_httponly_session_cookie_flask_login_final_filtered_merged_list.txt');
 // findInterestingRepos("Cookie-name-prefixes", "ut_session_cookie_name_manually_set.txt", true, 0, Number.MAX_VALUE, './repos_with_interesting_results/9bis - repos_with_manually_set_session_cookie_name_flask_login_final_filtered_merged_list.txt');
@@ -712,7 +712,7 @@ function libraryUsagesGrep() {
 // findOverlappingResultsInRepos({"Password-strength": ["un_using_django_built_in_forms.txt"], "Account-deactivation": ["un_custom_auth_backends.txt"]}, [true, true], './repos_with_interesting_results/18 - repos_that_use_a_custom_auth_backend_django_filtered_list_final_v2.txt');
 // findOverlappingResultsInRepos({"Password-strength": ["un_using_django_built_in_forms.txt"], "Logout-function-is-called": ["un_logout_function_is_called.txt"]}, [true, false], './repos_with_interesting_results/18 - repos_with_no_logout_django_filtered_list_final_v2.txt');
 // findOverlappingResultsInRepos({"Password-strength": ["un_using_django_built_in_forms.txt"], "Login-restrictions": ["un_no_authentication_checks_general.txt"]}, [true, true], './repos_with_interesting_results/18 - repos_with_no_auth_checks_django_filtered_list_final_v2.txt');
-deleteQueriesResults({"Password-strength": ["un_using_django_built_in_forms"]});
+// deleteQueriesResults({"Password-strength": ["un_using_django_built_in_forms"]});
 // deleteQueriesResults({".": ["django_library_used_check"]});
 // libraryUsagesGrep();
 // listMostCommonKeywordsAndUsers();
